BUg fixed Export
